refactor(project): new parser signature, migrate unrelated funcs
